# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a group of API for WTM Service",
        "title": "WTM Backend Service",
        "contact": {},
        "version": "1"
    },
    "paths": {
...
        "/bookings/booking-status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of booking statuses.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "List Booking Statuses",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of booking statuses",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.StatusBooking"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the status of a booking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Update booking status",
                "parameters": [
                    {
                        "description": "Update status booking request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bookingdto.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated booking status",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/bookings/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Finalize the cart by submitting guest data and changing booking status to \"in review\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Checkout cart",
                "responses": {
                    "200": {
                        "description": "Successfully checked out cart",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/bookingdto.DataInvoice"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/bookings/ids": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List booking IDs with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "List Booking IDs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved booking IDs",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseWithPagination"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/bookings/payment-status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of payment statuses.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "List Payment Statuses",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of payment statuses",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.StatusPayment"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the status of a payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Update payment status",
                "parameters": [
                    {
                        "description": "Update status payment request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bookingdto.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated payment status",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/bookings/receipt": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a receipt for a booking",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Upload Receipt",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Receipt File",
                        "name": "receipt",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "booking_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Sub Booking ID",
                        "name": "sub_booking_id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully uploaded receipt",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/bookings/{sub_booking_id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel an existing booking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Cancel Booking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "sub_booking_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully cancelled booking",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bookingdto.DataInvoice": {
            "type": "object",
            "properties": {
                "agent": {
                    "type": "string"
                },
                "check_in": {
                    "type": "string"
                },
                "check_out": {
                    "type": "string"
                },
                "company_agent": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "description_invoice": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DescriptionInvoice"
                    }
                },
                "email": {
                    "type": "string"
                },
                "guest": {
                    "type": "string"
                },
                "hotel": {
                    "type": "string"
                },
                "invoice_date": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "promo": {
                    "$ref": "#/definitions/entity.DetailPromo"
                },
                "receipt": {
                    "type": "string"
                },
                "sub_booking_id": {
                    "type": "string"
                },
                "total_price": {
                    "type": "number"
                }
            }
        },
...
        "bookingdto.UpdateStatusRequest": {
            "type": "object",
            "properties": {
                "booking_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status_id": {
                    "type": "integer"
                },
                "sub_booking_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.DescriptionInvoice": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "total": {
                    "type": "number"
                },
                "total_before_promo": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "entity.DetailPromo": {
            "type": "object",
            "properties": {
                "benefit_note": {
                    "type": "string"
                },
                "discount_percent": {
                    "type": "number"
                },
                "fixed_price": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "promo_code": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "upgraded_to_id": {
                    "type": "integer"
                }
            }
        },
...
        "entity.StatusBooking": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "entity.StatusPayment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.ResponseWithData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.ResponseWithPagination": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/response.Pagination"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}